+++ src/Network/Haskoin/Transaction.hs
@@ -10,8 +10,15 @@ Transactions and related code.
 -}
 module Network.Haskoin.Transaction
     ( module Common
+    -- * Builder
     , module Builder
+    -- * Segwit
+    , module Segwit
+    -- * Partial
+    , module Partial
     ) where
 
 import           Network.Haskoin.Transaction.Builder as Builder
 import           Network.Haskoin.Transaction.Common  as Common
+import           Network.Haskoin.Transaction.Partial as Partial
+import           Network.Haskoin.Transaction.Segwit  as Segwit